+++ src/lib.rs
@@ -231,7 +231,10 @@ pub use eval::EvalContext;
 #[cfg(not(feature = "no_object"))]
 use func::calc_typed_method_hash;
 use func::{calc_fn_hash, calc_fn_hash_full, calc_var_hash};
-pub use func::{plugin, FuncArgs, NativeCallContext, RhaiNativeFunc};
+pub use func::{
+    plugin, BorrowCallContext, BorrowedFuncArgs, BorrowedOpaquePointer, BorrowedScopeEntry,
+    BorrowedScopeValue, FuncArgs, NativeCallContext, RhaiNativeFunc,
+};
 pub use module::{FnNamespace, FuncRegistration, Module};
 pub use packages::string_basic::{FUNC_TO_DEBUG, FUNC_TO_STRING};
 pub use rhai_codegen::*;